requirements.txt
gspread>=6.0.0
google-auth>=2.0.0
requests>=2.31.0
python-dateutil>=2.8.0
